requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.12.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
nltk==3.6.7
numpy==1.22.1
oauthlib==3.1.1
onnx==1.10.2
onnxruntime==1.10.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.47
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tokenizers==0.10.3
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
zipp==3.7.0
